# Conflicts:
#	deployment/go.mod
#	deployment/go.sum
#	devenv/go.mod
#	devenv/go.sum
#	integration-tests/go.mod
#	integration-tests/go.sum
#	staging-monitor/go.mod

+++ deployment/ccip/config/deploy.go
@@ -52,7 +52,7 @@ type FeeQuoterParams struct {
 	ContractsSemver              *semver.Version
 	Coin                         string
 	MaxFeeJuelsPerMsg            *big.Int
-	TokenPriceStalenessThreshold uint64
+	TokenPriceStalenessThreshold uint32
 	FeeTokens                    map[TokenSymbol]FeeToken
 }
